experimentation pipeline/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.3.0
tqdm>=4.60.0
tensorboard>=2.5.0
PyWavelets>=1.1.1
PyYAML>=5.4.0
